requirements.txt
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
